@@ -1,12 +1,15 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using JoysOfEfficiency.Core;
 using JoysOfEfficiency.Utils;
 using Microsoft.Xna.Framework;
-using StardewModdingAPI;
 using StardewValley;
+using StardewValley.ItemTypeDefinitions;
 using StardewValley.Menus;
 using StardewValley.Objects;
+using StardewValley.SpecialOrders;
 using StardewValley.Tools;
+using SVObject = StardewValley.Object;
 
 namespace JoysOfEfficiency.Automation
 {
@@ -14,15 +17,16 @@ namespace JoysOfEfficiency.Automation
     {
         private static Config Config => InstanceHolder.Config;
 
-        private static readonly Logger Logger = new Logger("AFKFisher");
+        private static readonly Logger Logger = new Logger("AutoFisher");
 
         public static bool AfkMode { get; private set; }
+
+        public static int FishQuality { get; set; }
+        public static bool Treasure { get; set; }
 
         private static bool CatchingTreasure { get; set; }
         private static int AutoFishingCounter { get; set; }
         private static int AfkCooltimeCounter { get; set; }
-
-        private static IReflectionHelper Reflection => InstanceHolder.Reflection;
 
         public static void AfkFishing()
         {
@@ -74,9 +78,8 @@ namespace JoysOfEfficiency.Automation
             {
                 return;
             }
-            int whichFish = Reflection.GetField<int>(rod, "whichFish").GetValue();
 
-            if (!rod.isNibbling || !rod.isFishing || whichFish != -1 || rod.isReeling || rod.hit ||
+            if (!rod.isNibbling || !rod.isFishing || rod.whichFish != null || rod.isReeling || rod.hit ||
                 rod.isTimingCast || rod.pullingOutOfWater || rod.fishCaught || rod.castedButBobberStillInAir)
             {
                 return;
@@ -85,31 +88,18 @@ namespace JoysOfEfficiency.Automation
             rod.DoFunction(player.currentLocation, 1, 1, 1, player);
         }
 
-        public static void CollectFish(Farmer who, FishingRod rod)
+        private static void CollectFish(Farmer who, FishingRod rod)
         {
-            IReflectedField<int> recastTimerMs = Reflection.GetField<int>(rod, "recastTimerMs");
-
-            int whichFish = Reflection.GetField<int>(rod, "whichFish").GetValue();
-            int fishQuality = Reflection.GetField<int>(rod, "fishQuality").GetValue();
-            
-            string itemCategory = Reflection.GetField<string>(rod, "itemCategory").GetValue();
+            ItemMetadata whichFish = rod.whichFish;
+            String whichFishName = rod.whichFish.QualifiedItemId;
+            FishQuality = rod.fishQuality;
+            string fishID = whichFish.GetParsedData().ItemId;
+            int itemCategory = whichFish.GetParsedData().Category;
 
             if (!Game1.isFestival())
             {
                 who.faceDirection(2);
                 who.FarmerSprite.setCurrentFrame(84);
-            }
-
-            if (Game1.random.NextDouble() < 0.025)
-            {
-                who.currentLocation.temporarySprites.Add(new TemporaryAnimatedSprite("LooseSprites\\Cursors",
-                    new Rectangle(653, 858, 1, 1), 9999f, 1, 1,
-                    who.Position + new Vector2(Game1.random.Next(-3, 2) * 4, -32f), false, false,
-                    (float) (who.getStandingY() / 10000.0 + 1.0 / 500.0), 0.04f, Color.LightBlue, 5f, 0.0f,
-                    0.0f, 0.0f)
-                {
-                    acceleration = new Vector2(0.0f, 0.25f)
-                });
             }
 
             if (!who.IsLocalPlayer)
@@ -120,42 +110,47 @@ namespace JoysOfEfficiency.Automation
             who.currentLocation.localSound("coin");
             if (!rod.treasureCaught)
             {
-                recastTimerMs.SetValue(200);
-                Object @object = null;
+                SVObject @object = null;
                 switch (itemCategory)
                 {
-                    case "Object":
+                    case SVObject.furnitureCategory:
                     {
-                        @object = new Object(whichFish, 1, false, -1, fishQuality);
-                        if (whichFish == GameLocation.CAROLINES_NECKLACE_ITEM)
-                        {
-                            @object.questItem.Value = true;
-                        }
-
-                        if (whichFish == 79 || whichFish == 842)
-                        {
-                            @object = who.currentLocation.tryToCreateUnseenSecretNote(who);
-                            if (@object == null)
-                                return;
-                        }
-
-                        if (rod.caughtDoubleFish)
-                        {
-                            @object.Stack = 2;
-                        }
-
+                        Logger.Log($"\tFurniture? {fishID}\tItemCategory: {itemCategory}");
+                        @object = new Furniture(fishID, Vector2.Zero);
                         break;
                     }
-                    case "Furniture":
+                    case SVObject.junkCategory:
+                    case SVObject.litterCategory:
+                    case SVObject.FishCategory:
+                    default:
                     {
-                        @object = new Furniture(whichFish, Vector2.Zero);
+                        Logger.Log($"\tFishy, Litter, or Junky!  {fishID}");
+                        @object = new SVObject(fishID, 1, false, -1, FishQuality);
+                        if (fishID == GameLocation.CAROLINES_NECKLACE_ITEM_QID)
+                        {
+                            @object.questItem.Value = true;
+                            break;
+                        }
+
+                        if (fishID == "79" || fishID == "842") // Secret Note (79) or Journal Scrap (842)
+                        {
+                            @object = who.currentLocation.tryToCreateUnseenSecretNote(who);
+                            if (@object == null) return;
+                        }
+
+                        if (rod.numberOfFishCaught > 1)
+                        {
+                            @object.Stack = rod.numberOfFishCaught;
+                        }
+
                         break;
                     }
                 }
+
                 bool fromFishPond = rod.fromFishPond;
                 who.completelyStopAnimatingOrDoingAction();
                 rod.doneFishing(who, !fromFishPond);
-                if (!Game1.isFestival() && !fromFishPond && (itemCategory == "Object" && Game1.player.team.specialOrders.Count > 0))
+                if (!Game1.isFestival() && !fromFishPond && (itemCategory == SVObject.FishCategory && Game1.player.team.specialOrders.Count > 0))
                 {
                     foreach (SpecialOrder specialOrder in Game1.player.team.specialOrders)
                     {
@@ -168,40 +163,31 @@ namespace JoysOfEfficiency.Automation
                     return;
                 }
 
-                Game1.activeClickableMenu = new ItemGrabMenu(new List<Item>
-                {
-                    @object
-                }, rod).setEssential(true);
+                Game1.activeClickableMenu = new ItemGrabMenu(new List<Item> { @object }, rod).setEssential(true);
             }
             else
             {
+                Logger.Log($"Treazhure Cot!");
                 rod.fishCaught = false;
                 rod.showingTreasure = true;
                 who.UsingTool = true;
                 int initialStack = 1;
-                if (rod.caughtDoubleFish)
+                if (rod.numberOfFishCaught > 1)
                 {
-                    initialStack = 2;
+                    initialStack = rod.numberOfFishCaught;
                 }
 
-                Object @object = new Object(whichFish, initialStack, false, -1, fishQuality);
+                SVObject @object = new SVObject(fishID, initialStack, false, -1, FishQuality);
                 if (Game1.player.team.specialOrders.Count > 0)
                 {
+                    Logger.Log($"\tSpechul Treazhure!");
                     foreach (SpecialOrder specialOrder in Game1.player.team.specialOrders)
                     {
                         specialOrder.onFishCaught?.Invoke(Game1.player, @object);
                     }
                 }
                 bool inventoryBool = who.addItemToInventoryBool(@object);
-                rod.animations.Add(new TemporaryAnimatedSprite("LooseSprites\\Cursors", new Rectangle(64, 1920, 32, 32), 500f, 1, 0, who.Position + new Vector2(-32f, -160f), false, false, (float)(who.getStandingY() / 10000.0 + 1.0 / 1000.0), 0.0f, Color.White, 4f, 0.0f, 0.0f, 0.0f)
-                {
-                    motion = new Vector2(0.0f, -0.128f),
-                    timeBasedMotion = true,
-                    endFunction = rod.openChestEndFunction,
-                    extraInfoForEndBehavior = inventoryBool ? 0 : 1,
-                    alpha = 0.0f,
-                    alphaFade = -1f / 500f
-                });
+                rod.openChestEndFunction(inventoryBool ? 0 : 1);
             }
         }
 
@@ -213,22 +199,17 @@ namespace JoysOfEfficiency.Automation
                 return;
             }
 
-
-            IReflectedField<float> bobberSpeed = Reflection.GetField<float>(bar, "bobberBarSpeed");
-
-            float barPos = Reflection.GetField<float>(bar, "bobberBarPos").GetValue();
-            int barHeight = Reflection.GetField<int>(bar, "bobberBarHeight").GetValue();
-            float fishPos = Reflection.GetField<float>(bar, "bobberPosition").GetValue();
-            float treasurePos = Reflection.GetField<float>(bar, "treasurePosition").GetValue();
-            float distanceFromCatching = Reflection.GetField<float>(bar, "distanceFromCatching").GetValue();
-            bool treasureCaught = Reflection.GetField<bool>(bar, "treasureCaught").GetValue();
-            bool treasure = Reflection.GetField<bool>(bar, "treasure").GetValue();
-            float treasureAppearTimer = Reflection.GetField<float>(bar, "treasureAppearTimer").GetValue();
-            float bobberBarSpeed = bobberSpeed.GetValue();
-
+            float barPos = bar.bobberBarPos;
+            int barHeight = bar.bobberBarHeight;
+            float fishPos = bar.bobberPosition;
+            float treasurePos = bar.treasurePosition;
+            float distanceFromCatching = bar.distanceFromCatching;
+            bool treasureCaught = bar.treasureCaught;
+            float treasureAppearTimer = bar.treasureAppearTimer;
+            float bobberBarSpeed = bar.bobberBarSpeed;
             float top = barPos;
 
-            if (treasure && treasureAppearTimer <= 0 && !treasureCaught)
+            if (Treasure && treasureAppearTimer <= 0 && !treasureCaught)
             {
                 if (!CatchingTreasure && distanceFromCatching > 0.7f)
                 {
@@ -258,7 +239,7 @@ namespace JoysOfEfficiency.Automation
                 bobberBarSpeed = strength;
             }
 
-            bobberSpeed.SetValue(bobberBarSpeed);
+            bar.bobberBarSpeed = bobberBarSpeed;
         }
 
         public static void ToggleAfkFishing()
