@@ -12,13 +12,16 @@ namespace JoysOfEfficiency.Automation
 {
     internal class MachineOperator
     {
+        private static readonly Logger Logger = new Logger("MachineOperator");
+
         public static void DepositIngredientsToMachines()
         {
             Farmer player = Game1.player;
-            if (player.CurrentItem == null || !(Game1.player.CurrentItem is SVObject item))
+            if (player.CurrentItem == null || !(player.CurrentItem is SVObject item))
             {
                 return;
             }
+
             foreach (SVObject obj in Util.GetObjectsWithin<SVObject>(InstanceHolder.Config.MachineRadius).Where(IsObjectMachine))
             {
                 Vector2 loc = Util.GetLocationOf(currentLocation, obj);
@@ -28,11 +31,12 @@ namespace JoysOfEfficiency.Automation
                 if (obj.Name == "Keg" && item.ParentSheetIndex == 433 && item.Stack < 5)
                 {
                     // You don't have enough beans.
+                    Logger.Log($"Trying to deposit {item.Name} into KEG: {obj.Name}. Not enough beans!");
                     return;
                 }
 
-                bool flag = false;
                 bool accepted = obj.Name == "Furnace" ? CanFurnaceAcceptItem(item, player) : Utility.isThereAnObjectHereWhichAcceptsThisItem(currentLocation, item, (int)loc.X * tileSize, (int)loc.Y * tileSize);
+                Logger.Log($"Trying to deposit ({accepted}) {item.Name} into machine: {obj.Name}");
                 if (obj is Cask)
                 {
                     if (ModEntry.IsCoGOn || ModEntry.IsCaOn)
@@ -40,38 +44,39 @@ namespace JoysOfEfficiency.Automation
                         if (obj.performObjectDropInAction(item, true, player))
                         {
                             obj.heldObject.Value = null;
-                            flag = true;
+                            accepted = true;
                         }
                     }
-                    else if (currentLocation is Cellar && accepted)
+                    else if (currentLocation is not Cellar && accepted)
                     {
-                        flag = true;
+                        accepted = false;
                     }
                 }
-                else if (obj.Name == "Seed Maker")
+                else if (obj.Name == "Crab Pot")
                 {
-                    if (InstanceHolder.Config.AutoDepositSeedMaker == false)
+                    if (item.Name == "Bait" || item.Name == "Magic Bait")
                     {
-                        flag = false;
-                    }
-                    else
-                    {
-                        flag = true;
+                        accepted = true;
+                        Logger.Log($"\tCrab Pot and {item.Name} are now ACCEPTED.");
                     }
                 }
-                else if (accepted)
+                else if (obj.Name == "Seed Maker" && InstanceHolder.Config.AutoDepositSeedMaker == false)
                 {
-                    flag = true;
+                    continue;
                 }
 
-                if (!flag)
+                if (!accepted)
                     continue;
 
-                obj.performObjectDropInAction(item, false, player);
-                if (!(obj.Name == "Furnace" || obj.Name == "Charcoal Kiln") || item.Stack == 0)
+                // performObjectDropInAction but only if it's currently empty
+                if (obj.performObjectDropInAction(item, false, player, true))
                 {
                     player.reduceActiveItemByOne();
+                    Logger.Log($"Item {obj} MANUALLY consuming {item.Name}");
+                } else {
+                    Logger.Log($"Item {obj} should have already consumed {item.Name}");
                 }
+                Logger.Log($"DONE dropping {item.Name} into {obj}.");
 
                 return;
             }
@@ -82,8 +87,16 @@ namespace JoysOfEfficiency.Automation
             Farmer player = Game1.player;
             foreach (SVObject obj in Util.GetObjectsWithin<SVObject>(InstanceHolder.Config.MachineRadius).Where(IsObjectMachine))
             {
-                if (!obj.readyForHarvest.Value || obj.heldObject.Value == null)
+                // Nothing in the machine...
+                if (obj.heldObject.Value == null)
+                {
                     continue;
+                }
+                else if (!obj.readyForHarvest.Value)
+                {
+                    Logger.Log($"Time until {obj.Name} ready for collecting item {obj.heldObject.Value.Name}: {obj.MinutesUntilReady} game minutes.");
+                    continue;
+                }
 
                 Item item = obj.heldObject.Value;
                 if (player.couldInventoryAcceptThisItem(item))
@@ -93,64 +106,43 @@ namespace JoysOfEfficiency.Automation
 
         private static bool CanFurnaceAcceptItem(Item item, Farmer player)
         {
-            if (player.getTallyOfObject(382, false) <= 0)
+            Logger.Log($"{player.Items.ContainsId(Object.coalQID)} ** {item.Stack} ** {item.ParentSheetIndex}");
+
+            // Minimum of one coal in inventory
+            if (! player.Items.ContainsId(Object.coalQID, 1))
                 return false;
-            if (item.Stack < 5 && item.ParentSheetIndex != 80 && item.ParentSheetIndex != 82 && item.ParentSheetIndex != 330)
-                return false;
-            switch (item.ParentSheetIndex)
+
+            switch (item.Name)
             {
-                case 378:       // Copper Ore
-                case 380:       // Iron Ore
-                case 384:       // Gold Ore
-                case 386:       // Iridium Ore
-                case 80:        // Quartz
-                case 82:        // Fire Quartz
+                // One item per coal...
+                case "Clay":
+                case "Quartz":
+                case "Fire Quartz":
                     break;
+
+                // Five items per coal...
+                case "Copper Ore":
+                case "Iron Ore":
+                case "Gold Ore":
+                case "Iridium Ore":
+                case "Radioactive Ore":
+                    if (item.Stack < 5)
+                        return false;
+                    break;
+
                 default:
                     return false;
             }
             return true;
         }
 
-
-
-
         private static bool IsObjectMachine(SVObject obj)
         {
-            if (obj is CrabPot)
-                return true;
-
-            if (!obj.bigCraftable.Value)
-                return false;
-
-            switch (obj.Name)
+            if (InstanceHolder.Config.MachineTypes.Contains(obj.Name))
             {
-                case "Incubator":
-                case "Slime Incubator":
-                case "Keg":
-                case "Preserves Jar":
-                case "Cheese Press":
-                case "Mayonnaise Machine":
-                case "Loom":
-                case "Oil Maker":
-                case "Seed Maker":
-                case "Crystalarium":
-                case "Recycling Machine":
-                case "Furnace":
-                case "Charcoal Kiln":
-                case "Slime Egg-Press":
-                case "Cask":
-                case "Bee House":
-                case "Mushroom Box":
-                case "Statue Of Endless Fortune":
-                case "Statue Of Perfection":
-                case "Tapper":
-                case "Heavy Tapper":
-                case "Coffee Maker":
-                case "Worm Bin":
-                    return true;
-                default: return false;
+                return true;
             }
+            return false;
         }
     }
 }
