@@ -14,15 +14,14 @@ namespace JoysOfEfficiency.Core
         public static Config Config { get; private set; }
         private static IModHelper Helper => ModInstance.Helper;
         public static ITranslationHelper Translation => Helper.Translation;
-        public static IReflectionHelper Reflection => Helper.Reflection;
-        public static Multiplayer Multiplayer => Reflection.GetField<Multiplayer>(typeof(Game1), "multiplayer").GetValue();
+        public static Multiplayer Multiplayer => Game1.Multiplayer;
 
-        public static InputState Input => Reflection.GetField<InputState>(typeof(Game1), "input").GetValue();
+        public static InputState Input => Game1.input;
 
         public static CustomAnimalConfigHolder CustomAnimalTool;
 
         /// <summary>
-        /// Sets mod's entry　point and configuration instance. 
+        /// Sets mod's entry point and configuration instance.
         /// </summary>
         /// <param name="modInstance">the mod instance</param>
         /// <param name="conf">the configuration instance</param>
@@ -32,7 +31,7 @@ namespace JoysOfEfficiency.Core
             Config = LoadConfig();
             CustomAnimalTool = new CustomAnimalConfigHolder(modInstance.GetFilePath("customAnimalTools.json"));
         }
-        
+
         /// <summary>
         /// Writes settings to '(ModFolder)/config.json'.
         /// </summary>
