@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
 using JoysOfEfficiency.Core;
 using JoysOfEfficiency.OptionsElements;
 using JoysOfEfficiency.Utils;
@@ -86,6 +85,7 @@ namespace JoysOfEfficiency.Menus
                 tab.AddOptionsElement(new EmptyLabel());
                 tab.AddOptionsElement(new LabelComponent("Balanced Mode"));
                 tab.AddOptionsElement(new ModifiedCheckBox("BalancedMode", 20, Config.BalancedMode, OnCheckboxValueChanged));
+                tab.AddOptionsElement(new ModifiedSlider("EveryNthTick", 20, (int)Config.RunEveryNthTick, 5, 60, OnSliderValueChanged, () => Config.BalancedMode));
 
                 tab.AddOptionsElement(new EmptyLabel());
                 tab.AddOptionsElement(new LabelComponent("Auto Water Nearby Crops"));
@@ -109,7 +109,7 @@ namespace JoysOfEfficiency.Menus
                 tab.AddOptionsElement(new ModifiedSlider("CPUThresholdFishing", 0, (int)(Config.CpuThresholdFishing * 10), 0, 5, OnSliderValueChanged, () => !Config.AutoFishing, Format));
                 tab.AddOptionsElement(new ModifiedCheckBox("AutoReelRod", 6, Config.AutoReelRod, OnCheckboxValueChanged));
                 tab.AddOptionsElement(new ModifiedSlider("ThrowPower", 17, (int)(Config.ThrowPower * 10), 0, 10, OnSliderValueChanged, null, Format));
-                tab.AddOptionsElement(new ModifiedSlider("ThresholdStaminaPersentage", 18, Config.ThresholdStaminaPercentage, 10, 60, OnSliderValueChanged, null, Format));
+                tab.AddOptionsElement(new ModifiedSlider("ThresholdStaminaPercentage", 18, Config.ThresholdStaminaPercentage, 10, 60, OnSliderValueChanged, null, Format));
 
                 tab.AddOptionsElement(new EmptyLabel());
                 tab.AddOptionsElement(new LabelComponent("Auto Gate"));
@@ -120,6 +120,7 @@ namespace JoysOfEfficiency.Menus
                 tab.AddOptionsElement(new ModifiedCheckBox("AutoEat", 10, Config.AutoEat, OnCheckboxValueChanged));
                 tab.AddOptionsElement(new ModifiedSlider("StaminaToEatRatio", 1, (int)(Config.StaminaToEatRatio * 10), 1, 8, OnSliderValueChanged, () => !Config.AutoEat, Format));
                 tab.AddOptionsElement(new ModifiedSlider("HealthToEatRatio", 2, (int)(Config.HealthToEatRatio * 10), 1, 8, OnSliderValueChanged, () => !Config.AutoEat, Format));
+                tab.AddOptionsElement(new ModifiedCheckBox("DontEatThat", 45, Config.DontEatThat, OnCheckboxValueChanged));
 
                 tab.AddOptionsElement(new EmptyLabel());
                 tab.AddOptionsElement(new LabelComponent("Auto Harvest"));
@@ -163,17 +164,17 @@ namespace JoysOfEfficiency.Menus
                 tab.AddOptionsElement(new ModifiedCheckBox("AutoLootTreasures", 30, Config.AutoLootTreasures, OnCheckboxValueChanged));
                 tab.AddOptionsElement(new ModifiedCheckBox("CloseTreasureWhenAllLooted", 31, Config.CloseTreasureWhenAllLooted, OnCheckboxValueChanged));
 
-
                 tab.AddOptionsElement(new EmptyLabel());
                 tab.AddOptionsElement(new LabelComponent("Auto Pick Up Trash"));
                 tab.AddOptionsElement(new ModifiedCheckBox("AutoPickUpTrash", 34, Config.AutoPickUpTrash, OnCheckboxValueChanged));
+                tab.AddOptionsElement(new ModifiedCheckBox("GarbageDisgustsNPCs", 46, Config.GarbageDisgustsNPCs, OnCheckboxValueChanged));
                 tab.AddOptionsElement(new ModifiedSlider("ScavengingRadius", 13, Config.ScavengingRadius, 1, 3, OnSliderValueChanged, () => !Config.AutoPickUpTrash || Config.BalancedMode));
 
                 tab.AddOptionsElement(new EmptyLabel());
                 tab.AddOptionsElement(new LabelComponent("Auto Shearing and Milking"));
                 tab.AddOptionsElement(new ModifiedCheckBox("AutoShearingAndMilking", 35, Config.AutoShearingAndMilking, OnCheckboxValueChanged));
                 tab.AddOptionsElement(new ModifiedSlider("AnimalHarvestRadius", 14, Config.AnimalHarvestRadius, 1, 3, OnSliderValueChanged, () => !Config.AutoShearingAndMilking || Config.BalancedMode));
-                
+
                 tab.AddOptionsElement(new EmptyLabel());
                 tab.AddOptionsElement(new LabelComponent("Collect Letter Attachments And Quests"));
                 tab.AddOptionsElement(new ModifiedCheckBox("CollectLetterAttachmentsAndQuests", 36, Config.CollectLetterAttachmentsAndQuests, OnCheckboxValueChanged));
@@ -212,9 +213,16 @@ namespace JoysOfEfficiency.Menus
                 tab.AddOptionsElement(new EmptyLabel());
                 tab.AddOptionsElement(new LabelComponent("Fishing Probabilities Information"));
                 tab.AddOptionsElement(new ModifiedCheckBox("FishingProbabilitiesInfo", 26, Config.FishingProbabilitiesInfo, OnCheckboxValueChanged));
+                tab.AddOptionsElement(new ModifiedSlider("ProbBoxMaxFish", 19, Config.ProbBoxMaxFish, 5, 25, OnSliderValueChanged));
                 tab.AddOptionsElement(new ModifiedClickListener(this, "ProbBoxLocation", 0, Config.ProbBoxCoordinates.X, Config.ProbBoxCoordinates.Y, translation, OnSomewhereClicked, OnStartListeningClick));
                 tab.AddOptionsElement(new ModifiedCheckBox("MorePreciseProbabilities", 37, Config.MorePreciseProbabilities, OnCheckboxValueChanged, i => !Config.FishingProbabilitiesInfo));
                 tab.AddOptionsElement(new ModifiedSlider("TrialOfExamine", 15, Config.TrialOfExamine, 1, 50, OnSliderValueChanged, () => !(Config.FishingProbabilitiesInfo && Config.MorePreciseProbabilities)));
+
+                tab.AddOptionsElement(new EmptyLabel());
+                tab.AddOptionsElement(new LabelComponent("Fishing Tackle Information"));
+                tab.AddOptionsElement(new ModifiedCheckBox("FishingTackleInfo", 43, Config.FishingTackleInfo, OnCheckboxValueChanged));
+                tab.AddOptionsElement(new ModifiedCheckBox("TackleBoxAttach", 44, Config.TackleBoxAttach, OnCheckboxValueChanged));
+                tab.AddOptionsElement(new ModifiedClickListener(this, "TackleBoxLocation", 2, Config.TackleBoxCoordinates.X, Config.TackleBoxCoordinates.Y, translation, OnSomewhereClicked, OnStartListeningClick));
 
                 tab.AddOptionsElement(new EmptyLabel());
                 tab.AddOptionsElement(new LabelComponent("Show Shipping Price"));
@@ -296,6 +304,9 @@ namespace JoysOfEfficiency.Menus
                 case 1:
                     Config.PriceBoxCoordinates = point;
                     break;
+                case 2:
+                    Config.TackleBoxCoordinates = point;
+                    break;
                 default: return;
             }
             InstanceHolder.WriteConfig();
@@ -365,6 +376,10 @@ namespace JoysOfEfficiency.Menus
                 case 40: Config.BreakRocks = value; break;
                 case 41: Config.ChopTwigs = value; break;
                 case 42: Config.AutoDepositSeedMaker = value; break;
+                case 43: Config.FishingTackleInfo = value; break;
+                case 44: Config.TackleBoxAttach = value; break;
+                case 45: Config.DontEatThat = value; break;
+                case 46: Config.GarbageDisgustsNPCs = value; break;
                 default: return;
             }
             InstanceHolder.WriteConfig();
@@ -390,6 +405,8 @@ namespace JoysOfEfficiency.Menus
                 case 16: Config.RadiusFarmCleanup = value; break;
                 case 17: Config.ThrowPower = value / 10.0f; break;
                 case 18: Config.ThresholdStaminaPercentage = value; break;
+                case 19: Config.ProbBoxMaxFish = value; break;
+                case 20: Config.RunEveryNthTick = (uint)value; break;
                 default: return;
             }
 
@@ -852,7 +869,7 @@ namespace JoysOfEfficiency.Menus
         }
 
         /// <summary>
-        /// Try changing tab to a specified tab. 
+        /// Try changing tab to a specified tab.
         /// </summary>
         /// <param name="which">Which tab index to change.</param>
         private void TryToChangeTab(int which)
